refactor: modified header and navbar and more
